App_azure/requirements.txt
Flask>=1.0,<=1.1.2
requests
psycopg2-binary
